Major changes to the interpreter, lots of errors fixed and declaration made optional. Lexer updated to support floating point literals.

+++ expression_parser.py
@@ -40,7 +40,7 @@ def identifier(token):
 
 def literal(token):
 	try:
-		x=int(token)
+		x=float(token)
 		ret=True
 	except ValueError:
 		ret=False
